fix: moving function to utilities
